requirements.txt
streamlit
numpy
pandas
torch
torchvision
Pillow
requests
scikit-learn==1.0
